Move types to the top level `ActiveRecord` namespace
`ActiveRecord::ConnectionAdapters::Type::Value` =>
`ActiveRecord::Type::Value`
