--- activerecord/lib/active_record/connection_adapters/type.rb
@@ -1,25 +1,0 @@
-require 'active_record/connection_adapters/type/numeric'
-require 'active_record/connection_adapters/type/time_value'
-require 'active_record/connection_adapters/type/value'
-
-require 'active_record/connection_adapters/type/binary'
-require 'active_record/connection_adapters/type/boolean'
-require 'active_record/connection_adapters/type/date'
-require 'active_record/connection_adapters/type/date_time'
-require 'active_record/connection_adapters/type/decimal'
-require 'active_record/connection_adapters/type/decimal_without_scale'
-require 'active_record/connection_adapters/type/float'
-require 'active_record/connection_adapters/type/integer'
-require 'active_record/connection_adapters/type/string'
-require 'active_record/connection_adapters/type/text'
-require 'active_record/connection_adapters/type/time'
-
-require 'active_record/connection_adapters/type/type_map'
-require 'active_record/connection_adapters/type/hash_lookup_type_map'
-
-module ActiveRecord
-  module ConnectionAdapters
-    module Type # :nodoc:
-    end
-  end
-end
