--- activerecord/lib/active_record/connection_adapters/type/decimal_without_scale.rb
@@ -1,13 +1,0 @@
-require 'active_record/connection_adapters/type/integer'
-
-module ActiveRecord
-  module ConnectionAdapters
-    module Type
-      class DecimalWithoutScale < Integer # :nodoc:
-        def type
-          :decimal
-        end
-      end
-    end
-  end
-end
