--- activerecord/lib/active_record/connection_adapters/type/numeric.rb
@@ -1,20 +1,0 @@
-module ActiveRecord
-  module ConnectionAdapters
-    module Type
-      module Numeric # :nodoc:
-        def number?
-          true
-        end
-
-        def type_cast_for_write(value)
-          case value
-          when true then 1
-          when false then 0
-          when ::String then value.presence
-          else super
-          end
-        end
-      end
-    end
-  end
-end
