--- activerecord/lib/active_record/connection_adapters/type/text.rb
@@ -1,13 +1,0 @@
-require 'active_record/connection_adapters/type/string'
-
-module ActiveRecord
-  module ConnectionAdapters
-    module Type
-      class Text < String # :nodoc:
-        def type
-          :text
-        end
-      end
-    end
-  end
-end
