+++ activerecord/lib/active_record/type/binary.rb
@@ -1,0 +1,17 @@
+module ActiveRecord
+  module Type
+    class Binary < Value # :nodoc:
+      def type
+        :binary
+      end
+
+      def binary?
+        true
+      end
+
+      def klass
+        ::String
+      end
+    end
+  end
+end
