+++ activerecord/lib/active_record/type/boolean.rb
@@ -1,0 +1,19 @@
+module ActiveRecord
+  module Type
+    class Boolean < Value # :nodoc:
+      def type
+        :boolean
+      end
+
+      private
+
+      def cast_value(value)
+        if value == ''
+          nil
+        else
+          ConnectionAdapters::Column::TRUE_VALUES.include?(value)
+        end
+      end
+    end
+  end
+end
