+++ activerecord/lib/active_record/type/decimal_without_scale.rb
@@ -1,0 +1,11 @@
+require 'active_record/type/integer'
+
+module ActiveRecord
+  module Type
+    class DecimalWithoutScale < Integer # :nodoc:
+      def type
+        :decimal
+      end
+    end
+  end
+end
