+++ activerecord/lib/active_record/type/integer.rb
@@ -1,0 +1,27 @@
+module ActiveRecord
+  module Type
+    class Integer < Value # :nodoc:
+      include Numeric
+
+      def type
+        :integer
+      end
+
+      def klass
+        ::Fixnum
+      end
+
+      alias type_cast_for_database type_cast
+
+      private
+
+      def cast_value(value)
+        case value
+        when true then 1
+        when false then 0
+        else value.to_i rescue nil
+        end
+      end
+    end
+  end
+end
