+++ activerecord/lib/active_record/type/text.rb
@@ -1,0 +1,11 @@
+require 'active_record/type/string'
+
+module ActiveRecord
+  module Type
+    class Text < String # :nodoc:
+      def type
+        :text
+      end
+    end
+  end
+end
